fixed sorting alignment
